backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
python-dotenv
openai>=1.0.0
numpy